Windows single-GPU CUDA build: bundle NCCL stub + toolchain fixes
Bypass the hard NCCL dependency added by #6138 (multi-GPU CUDA) so that
single-GPU CUDA builds work on Windows, where NCCL has no official
distribution. All upstream commits are kept; NCCL code compiles as-is.

- CMakeLists: find_package(NCCL REQUIRED) -> QUIET, fall back to a bundled
  stub static library (cmake/nccl-stub) exposed as NCCL::NCCL. Safe because
  every NCCL call site only runs when num_gpu > 1; the stub aborts loudly
  if ever invoked at runtime. Also add -DFMT_UNICODE=0 to CUDA flags (nvcc
  builds fmt without MSVC /utf-8).
- cmake/nccl-stub: minimal nccl.h subset + aborting implementations.
- nccl_gbdt.hpp: guard unused <pthread.h> include with !_WIN32.
- meta.h: const -> constexpr for device-code constants (required by nvcc 13.x).
- cuda_single_gpu_tree_learner.cpp: fix std::max(size_t, unsigned long)
  ambiguity on Win64 with newer MSVC.
- fmt submodule: replace U"" char32_t literal with an int array (nvcc EDG
  rejects large char32_t string literal values).

+++ src/treelearner/cuda/cuda_single_gpu_tree_learner.cpp
@@ -562,9 +562,11 @@ void CUDASingleGPUTreeLearner::AllocateBitset() {
         max_cat_num_bin = std::max(bin_mapper->num_bin(), max_cat_num_bin);
       }
     }
-    // std::max(..., 1UL) to avoid error in the case when there are NaN's in the categorical values
-    const size_t cuda_bitset_max_size = std::max(static_cast<size_t>((max_cat_value + 31) / 32), 1UL);
-    const size_t cuda_bitset_inner_max_size = std::max(static_cast<size_t>((max_cat_num_bin + 31) / 32), 1UL);
+    // std::max(..., 1) to avoid error in the case when there are NaN's in the categorical values;
+    // both operands cast to size_t because on Win64 unsigned long is 32-bit and newer
+    // MSVC rejects std::max with mismatched argument types
+    const size_t cuda_bitset_max_size = std::max(static_cast<size_t>((max_cat_value + 31) / 32), static_cast<size_t>(1));
+    const size_t cuda_bitset_inner_max_size = std::max(static_cast<size_t>((max_cat_num_bin + 31) / 32), static_cast<size_t>(1));
     AllocateCUDAMemory<uint32_t>(&cuda_bitset_, cuda_bitset_max_size, __FILE__, __LINE__);
     AllocateCUDAMemory<uint32_t>(&cuda_bitset_inner_, cuda_bitset_inner_max_size, __FILE__, __LINE__);
     const int max_cat_in_split = std::min(config_->max_cat_threshold, max_cat_num_bin / 2);
